Lift Room
Lift room now takes the player up to the 2nd floor.

2nd floor design is far from finished currently just a boxed room with a space for a door.

Plan to line the walls with Statues and have two in front of the door.

New bug, the raycast allows you to teleport through the door into the other room and vice versa (and on to the lift platform) without having to complete the puzzle (handy for debugging and troubleshooting)